Move api.js to a separate folder
